moduuli8/requirements.txt
geographiclib==2.0
geopy==2.4.1
mysql-connector-python==9.2.0
